Move copier.compressGoroutine to compression.go
Only moves unchanged code, should not change behavior.
